viewport inputs

+++ AGE/src/Age/Renderer/OrthographicCameraController.cpp
@@ -24,19 +24,19 @@ namespace AGE {
     AGE_PROFILE_FUNCTION();
 
     if (Input::IsKeyPressed(Key::A)) {
-      m_CameraPosition.x -= cos(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
-      m_CameraPosition.y -= sin(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
+      m_CameraPosition.x -= cosf(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
+      m_CameraPosition.y -= sinf(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
     } else if (Input::IsKeyPressed(Key::D)) {
-      m_CameraPosition.x += cos(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
-      m_CameraPosition.y += sin(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
+      m_CameraPosition.x += cosf(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
+      m_CameraPosition.y += sinf(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
     }
 
     if (Input::IsKeyPressed(Key::W)) {
-      m_CameraPosition.x += -sin(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
-      m_CameraPosition.y += cos(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
+      m_CameraPosition.x += -sinf(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
+      m_CameraPosition.y += cosf(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
     } else if (Input::IsKeyPressed(Key::S)) {
-      m_CameraPosition.x -= -sin(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
-      m_CameraPosition.y -= cos(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
+      m_CameraPosition.x -= -sinf(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
+      m_CameraPosition.y -= cosf(glm::radians(m_CameraRotation)) * m_CameraTranslationSpeed * ts;
     }
 
     if (m_Rotation) {
